change client right and messages for data change

+++ backend/controllers/messages.js
@@ -3,15 +3,18 @@ import createError from "../utils/error.js";
 
 export const createMessage = async (req, res, next) => {
 
-    req.body.lastManipulatorId = req.client.id;
-    const newMessage = new Message(req.body);
-
     try {
-        const savedMessage = await newMessage.save();
-        res.status(200).json(savedMessage);
-        console.log(`${savedMessage.clientName} - ${savedMessage._id} - message has been saved!`);
-
+        if (req.body.messages != undefined && req.body.messages != "") {
+            req.body.lastManipulatorId = req.client.id;
+            const newMessage = new Message(req.body);
+            const savedMessage = await newMessage.save();
+            res.status(200).json(savedMessage);
+            console.log(`${savedMessage.clientName} - ${savedMessage._id} - message has been saved!`);
+        } else {
+            return next(createError(403, "Message is missing!"))
+        }
     }
+
     catch (error) {
         next(error);
     }
@@ -21,9 +24,14 @@ export const getMessages = async (req, res, next) => {
 
     try {
         const messages = (await Message.find()).filter((data) => data.isActive);
-        res.status(200).json(messages);
+        if (messages !== null) {
+            res.status(200).json(messages);
+        } else {
+            res.status(200).json("No messages in the database!");
+        }
+    }
 
-    } catch (error) {
+    catch (error) {
         next(error);
     }
 };
@@ -32,31 +40,46 @@ export const getMessageById = async (req, res, next) => {
 
     try {
         const actualMessage = await Message.findById(req.params.id);
-        res.status(200).json(actualMessage);
+        if (actualMessage !== null) {
+            res.status(200).json(actualMessage);
+        } else {
+            res.status(200).json("No message the given ID!");
+        }
+    }
 
-    } catch (error) {
+    catch (error) {
         next(error);
     }
 };
 
 export const updateMessageById = async (req, res, next) => {
-    const actualMessage = await Message.findById(req.params.id)
 
     try {
+        const actualMessage = await Message.findById(req.params.id)
         if (!req.client.isAdmin && !actualMessage.lastManipulatorId.match(req.client.id)) {
             return next(createError(403, "Not allowed to access that client data!"))
         }
         else {
-            req.body.lastManipulatorId = req.client.id;
-            const updatedMessage = await Message.findByIdAndUpdate(
-                req.params.id,
-                { $set: req.body },
-                { new: true }
-            );
-            res.status(200).json(updatedMessage);
-            console.log(`${updatedMessage._id} - ${updatedMessage.clientName} - message has been updated!`);
+            if (req.body.messages !== undefined || req.body.messages !== "") {
+                req.body.lastManipulatorId = req.client.id;
+                const updatedMessage = await Message.findByIdAndUpdate(
+                    req.params.id,
+                    { $set: req.body },
+                    { new: true }
+                );
+                if (updateAllergen !== null) {
+                    res.status(200).json(updatedMessage);
+                    console.log(`${updatedMessage._id} - ${updatedMessage.clientName} - message has been updated!`);
+                } else {
+                    res.status(200).json("No message the given ID!");
+                }
+            } else {
+                return next(createError(403, "Message is missing!"))
+            }
         }
-    } catch (error) {
+    }
+
+    catch (error) {
         next(error);
     }
 };
@@ -64,11 +87,17 @@ export const updateMessageById = async (req, res, next) => {
 export const deleteMessageById = async (req, res, next) => {
 
     try {
-        await Message.findByIdAndDelete(req.params.id);
-        res.status(200).json(`${req.params.id} - message has been deleted!`);
-        console.log(`${req.params.id} - message has been deleted!`);
+        const actualMessage = await Message.findById(req.params.id);
+        if (actualMessage !== null) {
+            const deletedMessage = await Message.findByIdAndDelete(req.params.id);
+            res.status(200).json(`${req.params.id} - message has been deleted!`);
+            console.log(`${req.params.id} - message has been deleted!`);
+        } else {
+            res.status(200).json("No message the given ID!");
+        }
+    }
 
-    } catch (error) {
+    catch (error) {
         next(error);
     }
 };
